Plane: rename arming_checks to AP_Arming
